requirements.txt
flatbuffers==2.0.0
matplotlib==3.7.1
opencv-contrib-python==4.8.1.78
psutil==5.9.4
pyautogui==0.9.53
customtkinter==5.2.0
mediapipe==0.9.3.0
PyDirectInput==1.0.4;  sys_platform == 'win32'
pywin32==306;  sys_platform == 'win32'
pygrabber;  sys_platform == 'win32'
pyinstaller==5.11.0
Pillow==10.1.0
numpy==1.26.2

# Embedded HTML for the Tk user interface system.
tkinterweb >= 3.23.3

pywebview >=5.0.5
